@@ -4,7 +4,7 @@ version: 1.0.1
 homepage: https://lbs.amap.com/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
@@ -12,6 +12,7 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter  
+  js: ^0.6.3
 
 dev_dependencies:
   flutter_test:
@@ -35,7 +36,7 @@ flutter:
         pluginClass: AMapFlutterLocationPlugin
       web: 
         pluginClass: AmapFlutterLocationWeb
-        fileName: amap_flutter_location_web.dart
+        fileName: ./web/amap_flutter_location_web.dart
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
