@@ -1,12 +1,16 @@
 import 'dart:async';
+import 'dart:collection';
+import 'dart:convert';
 // In order to *not* need this ignore, consider extracting the "web" version
 // of your plugin as a separate package, instead of inlining it in the same
 // package as the core of your plugin.
 // ignore: avoid_web_libraries_in_flutter
 import 'dart:html' as html show window;
 
+import 'package:amap_flutter_location/web/amap_js.dart' as AMap;
 import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+import 'package:js/js.dart';
 
 /// A web implementation of the AmapFlutterLocation plugin.
 class AmapFlutterLocationWeb {
@@ -26,7 +30,7 @@ class AmapFlutterLocationWeb {
   // static const String _CHANNEL_STREAM_LOCATION = "amap_flutter_location_stream";
   // static const EventChannel _eventChannel =
   //     const EventChannel(_CHANNEL_STREAM_LOCATION);
-  String _pluginKey;
+  late String _pluginKey;
   // static Stream<Map<String, Object>> _onLocationChanged = _eventChannel
   //     .receiveBroadcastStream()
   //     .asBroadcastStream()
@@ -45,8 +49,8 @@ class AmapFlutterLocationWeb {
         return getPlatformVersion();
       case 'setApiKey':
         return setApiKey();
-      case 'onLocationChanged':
-        return onLocationChanged();
+      case 'startLocation':
+        return startLocation();
       default:
         throw PlatformException(
           code: 'Unimplemented',
@@ -67,23 +71,24 @@ class AmapFlutterLocationWeb {
         'web端在index.html自行加<script type="text/javascript" src="https://webapi.amap.com/maps?v=1.4.15&key=您申请的key值"></script>');
   }
 
-  onLocationChanged() {
-    // if (_receiveStream == null) {
-    final receiveStream = StreamController<Map<String, Object>>();
-    //   _subscription = _onLocationChanged.listen((Map<String, Object> event) {
-    //     if (event != null && event['pluginKey'] == _pluginKey) {
-    //       Map<String, Object> newEvent = Map<String, Object>.of(event);
-    //       newEvent.remove('pluginKey');
-    //       _receiveStream.add(newEvent);
-    //     }
-    //   });
-    // }
-    receiveStream.add({'x': 1});
-    final x = receiveStream.stream;
-    print(x);
-    return x;
-  }
-
   ///开始定位
-  void startLocation() {}
+  Future<Map<String, Object>> startLocation() async {
+    AMap.plugin('AMap.Geolocation', () {
+      print(AMap.Geolocation().getCurrentPosition);
+      // final x =
+      //     AMap.Geolocation(AMap.GeolocationOptions(enableHighAccuracy: true));
+      // x.getCurrentPosition((status) {
+      //   print(1);
+      //   print(status);
+      // });
+      //AMap.Geolocation().getCurrentPosition((status, result) {});
+    });
+
+    return {};
+  }
+}
+
+class X {
+  X(this.x);
+  int x;
 }
